Rename composite fault specification.
